// src/frontend/react-admin/src/generated/models/role-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface RoleResponse
 */
export interface RoleResponse {
    /**
     * 
     * @type {number}
     * @memberof RoleResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RoleResponse
     */
    nom?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleResponse
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof RoleResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof RoleResponse
     */
    updatedAt?: Date;
}